Fix unnecessary encapsulation
